requirements.txt
pystray
plyer
pillow  # For handling the icon